flattens tests
